Updates to stability
Changed `loading` to a more descriptive `isWorking`. This makes is a bit more clear what is happening in the app. Change is primarily for other developers and code readability.

Removed some `useEffect` calls. This was done to avoid running that lifecycle hook when it wasn't needed. The guard clause of `isWorking` was a good protection, but it is good to avoid a call whenever possible.

Changed some string concatenation over to use template literals (widely supported 97%) as they get past the potential edge case of a string with two number next to each other. When using the `+=` operation there is a chance js might perform arithmetic instead of concatenation.

Changed the setting of isWorking to be more deliberate and tied to specific operations. This would make it easier to add other operations is the future that also make use of `isWorking` without the worry of collisions or operations being run when not needed.

Added an error reporter. In the case where some operation does not exist there will be a log of it displayed on the page along with some color change of the operation in question. The error display is limited to one but will show how many times the operation would have been performed. *Note: One error was left in a method name for the scramble option as a demo of this functionality.

Users can remove any operation from the list as long as work is not being done.

Ensured that only positive integers can be entered as a length of the random array.

Made sure all states are initialized with some value to avoid any unexpected bugs or strange ui behavior based on those items at startup.